Changes to be committed: 	modified:   BE/app/Http/Controllers/AuthController.php 	renamed:    BE/app/Http/Controllers/FilterController.php -> BE/app/Http/Controllers/DataController.php 	modified:   BE/app/Http/Controllers/DosenController.php 	modified:   BE/app/Http/Controllers/JadwalController.php 	modified:   BE/app/Http/Controllers/LaboranController.php 	modified:   BE/app/Http/Controllers/MakulController.php 	modified:   BE/app/Http/Controllers/PeriodeTahunController.php 	modified:   BE/app/Http/Controllers/ProdiController.php 	modified:   BE/app/Http/Controllers/RuanganController.php 	modified:   BE/app/Http/Controllers/WaktuPerkuliahanController.php 	new file:   BE/app/Http/Requests/RegisterUserRequest.php 	deleted:    BE/app/Http/Requests/StoreBebanAjarRequest.php 	modified:   BE/app/Http/Requests/StoreDosenRequest.php 	modified:   BE/app/Http/Requests/StoreJadwalRequest.php 	deleted:    BE/app/Http/Requests/UpdateBebanAjarRequest.php 	deleted:    BE/app/Rules/JadwalValidationRule.php 	new file:   BE/database/migrations/2025_12_08_125125_add_field_to_waktu_perkuliahans.php 	modified:   BE/routes/api.php

+++ BE/app/Http/Controllers/MakulController.php
@@ -57,7 +57,7 @@ class MakulController extends Controller implements HasMiddleware
             'nama_makul' => 'required'
         ]);//
 
-        $makul->Makul::update($fields);
+        $makul->update($fields);
         return ['makul' => $makul];//
         //
         //
@@ -68,6 +68,13 @@ class MakulController extends Controller implements HasMiddleware
      */
     public function destroy(Makul $makul)
     {
-        $makul->delete();//
+        if($makul->jadwal()->exists()){
+            return response("Data makul tidak bisa dihapus karena terdapat di data jadwal", 422);
+        }else{
+            $makul->delete();//
+            return[
+                "messsage"=>"Data makul berhasil dihapus"
+            ];
+        }
     }
 }
